Turn SourceForm into true container
